Fix absent types and mismatches

+++ common/src/main/kotlin/com/gitlab/kordlib/common/entity/Activity.kt
@@ -22,8 +22,8 @@ data class Activity(
 
 @Serializable
 data class ActivityTimeStamps(
-        val start: Int? = null,
-        val end: Int? = null
+        val start: Long? = null,
+        val end: Long? = null
 )
 
 @Serializable
